arm: FreeBSD/ARMv6 port
Sponsored by:	The FreeBSD Foundation
