+++ src/arm/Gstep.c
@@ -80,99 +80,6 @@ arm_exidx_step (struct cursor *c)
 }
 
 PROTECTED int
-unw_handle_signal_frame (unw_cursor_t *cursor)
-{
-  struct cursor *c = (struct cursor *) cursor;
-  int ret;
-  unw_word_t sc_addr, sp, sp_addr = c->dwarf.cfa;
-  struct dwarf_loc sp_loc = DWARF_LOC (sp_addr, 0);
-
-  if ((ret = dwarf_get (&c->dwarf, sp_loc, &sp)) < 0)
-    return -UNW_EUNSPEC;
-
-  /* Obtain signal frame type (non-RT or RT). */
-  ret = unw_is_signal_frame (cursor);
-
-  /* Save the SP and PC to be able to return execution at this point
-     later in time (unw_resume).  */
-  c->sigcontext_sp = c->dwarf.cfa;
-  c->sigcontext_pc = c->dwarf.ip;
-
-  /* Since kernel version 2.6.18 the non-RT signal frame starts with a
-     ucontext while the RT signal frame starts with a siginfo, followed
-     by a sigframe whose first element is an ucontext.
-     Prior 2.6.18 the non-RT signal frame starts with a sigcontext while
-     the RT signal frame starts with two pointers followed by a siginfo
-     and an ucontext. The first pointer points to the start of the siginfo
-     structure and the second one to the ucontext structure.  */
-
-  if (ret == 1)
-    {
-      /* Handle non-RT signal frames. Check if the first word on the stack
-         is the magic number.  */
-      if (sp == 0x5ac3c35a)
-        {
-          c->sigcontext_format = ARM_SCF_LINUX_SIGFRAME;
-          sc_addr = sp_addr + LINUX_UC_MCONTEXT_OFF;
-        }
-      else
-        {
-          c->sigcontext_format = ARM_SCF_LINUX_OLD_SIGFRAME;
-          sc_addr = sp_addr;
-        }
-    }
-  else if (ret == 2)
-    {
-      /* Handle RT signal frames. Check if the first word on the stack is a
-         pointer to the siginfo structure.  */
-      if (sp == sp_addr + 8)
-        {
-          c->sigcontext_format = ARM_SCF_LINUX_OLD_RT_SIGFRAME;
-          sc_addr = sp_addr + 8 + sizeof (siginfo_t) + LINUX_UC_MCONTEXT_OFF;
-        }
-      else
-        {
-          c->sigcontext_format = ARM_SCF_LINUX_RT_SIGFRAME;
-          sc_addr = sp_addr + sizeof (siginfo_t) + LINUX_UC_MCONTEXT_OFF;
-        }
-    }
-  else
-    return -UNW_EUNSPEC;
-
-  c->sigcontext_addr = sc_addr;
-  c->frame_info.frame_type = UNW_ARM_FRAME_SIGRETURN;
-  c->frame_info.cfa_reg_offset = sc_addr - sp_addr;
-
-  /* Update the dwarf cursor.
-     Set the location of the registers to the corresponding addresses of the
-     uc_mcontext / sigcontext structure contents.  */
-  c->dwarf.loc[UNW_ARM_R0] = DWARF_LOC (sc_addr + LINUX_SC_R0_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R1] = DWARF_LOC (sc_addr + LINUX_SC_R1_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R2] = DWARF_LOC (sc_addr + LINUX_SC_R2_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R3] = DWARF_LOC (sc_addr + LINUX_SC_R3_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R4] = DWARF_LOC (sc_addr + LINUX_SC_R4_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R5] = DWARF_LOC (sc_addr + LINUX_SC_R5_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R6] = DWARF_LOC (sc_addr + LINUX_SC_R6_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R7] = DWARF_LOC (sc_addr + LINUX_SC_R7_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R8] = DWARF_LOC (sc_addr + LINUX_SC_R8_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R9] = DWARF_LOC (sc_addr + LINUX_SC_R9_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R10] = DWARF_LOC (sc_addr + LINUX_SC_R10_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R11] = DWARF_LOC (sc_addr + LINUX_SC_FP_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R12] = DWARF_LOC (sc_addr + LINUX_SC_IP_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R13] = DWARF_LOC (sc_addr + LINUX_SC_SP_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R14] = DWARF_LOC (sc_addr + LINUX_SC_LR_OFF, 0);
-  c->dwarf.loc[UNW_ARM_R15] = DWARF_LOC (sc_addr + LINUX_SC_PC_OFF, 0);
-
-  /* Set SP/CFA and PC/IP.  */
-  dwarf_get (&c->dwarf, c->dwarf.loc[UNW_ARM_R13], &c->dwarf.cfa);
-  dwarf_get (&c->dwarf, c->dwarf.loc[UNW_ARM_R15], &c->dwarf.ip);
-
-  c->dwarf.pi_valid = 0;
-
-  return 1;
-}
-
-PROTECTED int
 unw_step (unw_cursor_t *cursor)
 {
   struct cursor *c = (struct cursor *) cursor;
@@ -216,14 +123,18 @@ unw_step (unw_cursor_t *cursor)
 
   /* Fall back on APCS frame parsing.
      Note: This won't work in case the ARM EABI is used. */
+#ifdef __FreeBSD__
+  if (0)
+#else
   if (unlikely (ret < 0))
+#endif
     {
       if (UNW_TRY_METHOD(UNW_ARM_METHOD_FRAME))
         {
+          Debug (13, "dwarf_step() failed (ret=%d), trying frame-chain\n", ret);
           ret = UNW_ESUCCESS;
           /* DWARF unwinding failed, try to follow APCS/optimized APCS frame chain */
           unw_word_t instr, i;
-          Debug (13, "dwarf_step() failed (ret=%d), trying frame-chain\n", ret);
           dwarf_loc_t ip_loc, fp_loc;
           unw_word_t frame;
           /* Mark all registers unsaved, since we don't know where
@@ -266,7 +177,7 @@ unw_step (unw_cursor_t *cursor)
               c->dwarf.loc[UNW_ARM_R12] = ip_loc;
               c->dwarf.loc[UNW_ARM_R11] = fp_loc;
               c->dwarf.pi_valid = 0;
-              Debug(15, "ip=%lx\n", c->dwarf.ip);
+              Debug(15, "ip=%x\n", c->dwarf.ip);
             }
           else
             {
